requirements.txt
Django==3.2.17
djangorestframework~=3.11.1
django-filter~=2.4.0
drf-yasg~=1.17.1
python-decouple==3.3
django-silk==4.1.0
psycopg2-binary==2.8.6
django-cors-headers==3.13.0
channels==3.0.1
channels-redis==4.0.0
django-redis==5.2.0
djangorestframework-jwt==1.11.0
